Get rid of ris references, improve error handling
